+++ packages/client/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.6.0
+* Expose `addActions` to enable extending the client with new actions that have the same access levels as first party ones.
+* Expose `addReactors` which allows extending the client with new reactors.
+* Expose `Tree`, `encode`, `decode`, `hasValue`, and `hasContext` utils.
+
 # 2.5.2
 * Fix defaults for facet and results
 * Add more reactors across many example types